third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/15.4.4.12-expected.txt

PASS 'use strict'; var a = arr(); [a.splice(0, 1), a] threw exception of type TypeError.
PASS var a = arr(); [a.splice(0, 1), a] threw exception of type TypeError.
PASS true === true
PASS 'use strict'; var o = obj(); [Array.prototype.splice.call(o, 0, 1), o] threw exception of type TypeError.
PASS var o = obj(); [Array.prototype.splice.call(o, 0, 1), o] threw exception of type TypeError.
PASS true === true
PASS 'use strict'; var a = agap(); [a.splice(0, 1), a] threw exception of type TypeError.
PASS var a = agap(); [a.splice(0, 1), a] threw exception of type TypeError.
PASS true === true
PASS 'use strict'; var o = ogap(); [Array.prototype.splice.call(o, 0, 1), o] threw exception of type TypeError.
PASS var o = ogap(); [Array.prototype.splice.call(o, 0, 1), o] threw exception of type TypeError.
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
